more image output formats
git-svn-id: https://pswww.slac.stanford.edu/svn-readonly/psdmrepo@3530 b967ad99-d558-0410-b138-e0f6c56caec7

+++ XtcExplorer/trunk/src/cspad.py
@@ -275,7 +275,7 @@ class CsPad( object ):
         """
         self.pixels = data2d.reshape(4,8,185,388)
 
-        self.image = np.zeros((2*self.npix_quad+100, 2*self.npix_quad+100 ), dtype="float64")
+        self.image = np.zeros((2*self.npix_quad+100, 2*self.npix_quad+100 ), dtype=self.pixels.dtype)
         for quad in xrange (4):
 
             quad_image = self.get_quad_image( self.pixels[quad], quad )
